fix: renaming types for clarity
