+++ src/onemod/types/integer.py
@@ -2,12 +2,12 @@ from typing_extensions import Annotated
 
 from pydantic import BaseModel, Field
 
-class BoundedInteger(BaseModel):
+class Integer(BaseModel):
     value: Annotated[int, Field]
     
     @classmethod
     def with_bounds(cls, ge: int | None = None, le: int | None = None) -> BaseModel:
-        """Factory method to create a custom bound BoundedInteger model"""
+        """Factory method to create a custom bounded Integer model"""
         field_args = {}
         if ge is None:
             field_args['le'] = le
